harmony_tasklist_manager/pyproject.toml
[project]
name = "harmony-tasklist-manager"
version = "1.0.0"
description = "HarmonyOS Task List Manager MCP Server"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp>=0.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
harmony-tasklist-server = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
